Move cmd/main.go code to cmd/juno.go
