+++ cmd/juno/juno.go
@@ -1,6 +1,12 @@
 package main
 
 import (
+	"context"
+	"fmt"
+	"os"
+	"os/signal"
+	"syscall"
+
 	"github.com/NethermindEth/juno/node"
 	"github.com/NethermindEth/juno/utils"
 	"github.com/mitchellh/mapstructure"
@@ -8,7 +14,15 @@ import (
 	"github.com/spf13/viper"
 )
 
-var Version string
+const greeting = `
+       _                    
+      | |                   
+      | |_   _ _ __   ___   
+  _   | | | | | '_ \ / _ \  
+ | |__| | |_| | | | | (_) |  
+  \____/ \__,_|_| |_|\___/  
+
+Juno is a Go implementation of a Starknet full node client created by Nethermind.`
 
 const (
 	configF   = "config"
@@ -31,6 +45,36 @@ const (
 	networkUsage = "Options: mainnet, goerli, goerli2, integration."
 	pprofUsage   = "Enables the pprof server and listens on port 9080."
 )
+
+var Version string
+
+func main() {
+	quit := make(chan os.Signal, 1)
+	signal.Notify(quit, os.Interrupt, syscall.SIGTERM)
+
+	ctx, cancel := context.WithCancel(context.Background())
+	go func() {
+		<-quit
+		cancel()
+	}()
+
+	config := new(node.Config)
+	cmd := NewCmd(config, func(cmd *cobra.Command, _ []string) error {
+		fmt.Printf("%s\n\n", greeting)
+
+		n, err := node.New(config)
+		if err != nil {
+			return err
+		}
+
+		n.Run(cmd.Context())
+		return nil
+	})
+
+	if err := cmd.ExecuteContext(ctx); err != nil {
+		os.Exit(1)
+	}
+}
 
 // NewCmd returns a command that can be exected with any of the Cobra Execute* functions.
 // The RunE field is set to the user-provided run function, allowing for robust testing setups.
