--- cmd/juno/main.go
@@ -1,50 +1,0 @@
-package main
-
-import (
-	"context"
-	"fmt"
-	"os"
-	"os/signal"
-	"syscall"
-
-	"github.com/NethermindEth/juno/node"
-	"github.com/spf13/cobra"
-)
-
-const greeting = `
-       _                    
-      | |                   
-      | |_   _ _ __   ___   
-  _   | | | | | '_ \ / _ \  
- | |__| | |_| | | | | (_) |  
-  \____/ \__,_|_| |_|\___/  
-
-Juno is a Go implementation of a Starknet full node client created by Nethermind.`
-
-func main() {
-	quit := make(chan os.Signal, 1)
-	signal.Notify(quit, os.Interrupt, syscall.SIGTERM)
-
-	ctx, cancel := context.WithCancel(context.Background())
-	go func() {
-		<-quit
-		cancel()
-	}()
-
-	config := new(node.Config)
-	cmd := NewCmd(config, func(cmd *cobra.Command, _ []string) error {
-		fmt.Printf("%s\n\n", greeting)
-
-		n, err := node.New(config)
-		if err != nil {
-			return err
-		}
-
-		n.Run(cmd.Context())
-		return nil
-	})
-
-	if err := cmd.ExecuteContext(ctx); err != nil {
-		os.Exit(1)
-	}
-}
